refactor: extract governance validation helpers to public storage.governance_validation
